requirements.txt
PySide6>=6.6.0
openai>=1.0.0,<2.0.0
azure-ai-documentintelligence>=1.0.0b2
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pydantic>=2.0.0 